requirements.txt
django
djangorestframework
django-cors-headers
pytube
pandas
openpyxl
python-dotenv